Move image detail modal to SearchResult component
