+++ src/app/pages/Search/Search.module.css
@@ -9,34 +9,6 @@
   margin: var(--gapSmall);
 }
 
-.searchResult {
-  height: 100%;
-  overflow-y: auto;
-  overflow-x: hidden;
-}
-
-.modalContent {
-  display: grid;
-  grid-template-rows: auto 6rem 1fr;
-  height: 100%;
-}
-
-.modalTabContent {
-  overflow-y: auto;
-  padding: 1rem;
-}
-
-.searchResult::-webkit-scrollbar,
-.modalTabContent::-webkit-scrollbar {
-  width: var(--gapSmall);
-}
-
-.searchResult::-webkit-scrollbar-thumb,
-.modalTabContent::-webkit-scrollbar-thumb {
-  background: var(--primary);
-  border-radius: var(--radius);
-}
-
 .popupSmall {
   display: grid;
   grid: 100% / auto 1fr auto;
