+++ src/app/pages/Search/Search.tsx
@@ -1,18 +1,13 @@
 import axios from 'axios';
 import React, { useEffect, useRef, useState } from 'react';
-import { isMobileOnly } from 'react-device-detect';
 import { useHistory, useParams } from 'react-router';
 import type { filtersAspectRatio, image } from '../../../lib/types/image';
 import Button from '../../components/Button/Button';
 import FilterBar from '../../components/FilterBar/FilterBar';
 import Headline from '../../components/Headline/Headline';
 import Icon from '../../components/Icon/Icon';
-import ImageDetails from '../../components/ImageDetails/ImageDetails';
 import Input from '../../components/Input/Input';
 import Modal from '../../components/Modal/Modal';
-import type { NavBarImageItems } from '../../components/NavBarImage/NavBarImage';
-import NavBarImage from '../../components/NavBarImage/NavBarImage';
-import PreviewImage from '../../components/PreviewImage/PreviewImage';
 import SearchResult from '../../components/SearchResult/SearchResult';
 import useCurrentUser from '../../hooks/useCurrentUser';
 import useFetchSearchImages from '../../hooks/useFetchSearchImages';
@@ -20,36 +15,6 @@ import styles from './Search.module.css';
 
 export type SearchProps = {
   className?: string;
-};
-
-const imageSize = {
-  desktop: {
-    maxHeight: 45,
-    maxWidth: 60,
-  },
-  mobile: {
-    maxHeight: 40,
-    maxWidth: 100,
-  },
-};
-
-const modalDetailsSize = {
-  desktop: {
-    minHeight: ``,
-    minWidth: ``,
-    height: `${2 * imageSize.desktop.maxHeight}%`,
-    width: `${imageSize.desktop.maxWidth}%`,
-    maxHeight: ``,
-    maxWidth: ``,
-  },
-  mobile: {
-    minHeight: ``,
-    minWidth: ``,
-    height: `${2.5 * imageSize.mobile.maxHeight}%`,
-    width: `${imageSize.mobile.maxWidth}%`,
-    maxHeight: ``,
-    maxWidth: ``,
-  },
 };
 
 const popupSize = {
@@ -78,9 +43,6 @@ export default function Search({ className = '' }: SearchProps): JSX.Element {
   const [searchValue, setSearchValue] = useState<string>('');
   const [fetchMoreImages, setFetchMoreImages] = useState<boolean>(false);
 
-  const [selectedImage, setSelectedImage] = useState<image | null>(null);
-  const [modalActiveTab, setModalActiveTab] = useState<NavBarImageItems>('details');
-
   const [filterAspectRatio, setFilterAspectRatio] = useState<filtersAspectRatio>('nofilter');
   const [filterColor, setFilterColor] = useState<number>(0);
   const [addToCollectionResult, setAddToCollectionResult] = useState<{
@@ -106,23 +68,12 @@ export default function Search({ className = '' }: SearchProps): JSX.Element {
   );
 
   useEffect(() => {
-    if (inputValue) handleSubmit();
-  });
+    if (query) handleSubmit();
+  }, [query]);
 
   useEffect(() => {
     setFetchMoreImages(false);
   }, [fetchMoreImages]);
-
-  function calcImageHeight(): number {
-    if (!selectedImage) return 0;
-
-    const imageMaxSize = isMobileOnly ? imageSize.mobile : imageSize.desktop;
-    const imageAspectRatio = selectedImage.height / selectedImage.width || 1;
-    return Math.min(
-      ((window.innerWidth * imageMaxSize.maxWidth) / 100) * imageAspectRatio,
-      (window.innerHeight * imageMaxSize.maxHeight) / 100
-    );
-  }
 
   function handleScroll(position: number, parentHeight: number) {
     if (!fetchMoreImages && 3 * parentHeight > -position) setFetchMoreImages(true);
@@ -169,8 +120,6 @@ export default function Search({ className = '' }: SearchProps): JSX.Element {
     }, 4000);
   }
 
-  const imageHeight = calcImageHeight();
-
   return (
     <>
       <main className={`${styles.search} ${className}`}>
@@ -193,70 +142,51 @@ export default function Search({ className = '' }: SearchProps): JSX.Element {
           isLoading={isLoading}
           isFetchingNewResult={isFetchingNewResult}
           imagesResult={imagesResult}
-          onImageClick={(index) => {
-            setSelectedImage(imagesResult?.results[index] || null);
-            setModalActiveTab('details');
-          }}
           onCollectionClick={(index) => handleAddToCollection(imagesResult?.results[index])}
           handleScroll={handleScroll}
           className={styles.searchResult}
         />
 
         <Modal
-          show={!!selectedImage}
-          backgroundBlur
-          closeButton={!isMobileOnly}
-          backButton={isMobileOnly}
-          onClose={() => setSelectedImage(null)}
-          size={modalDetailsSize}
+          show={modalAddToCollectionExtended === false}
+          position="bottomRightSlide"
+          size={popupSize}
+          backgroundOverlay={false}
         >
-          <div className={styles.modalContent}>
-            <div style={{ height: `${imageHeight}px` }}>
-              <PreviewImage image={selectedImage} />
-            </div>
-            <NavBarImage onClick={(item) => setModalActiveTab(item)} active={modalActiveTab} />
-            <div className={styles.modalTabContent}>
-              {modalActiveTab === 'details' && <ImageDetails image={selectedImage} />}
-              {modalActiveTab === 'collection' && 'collection'}
-              {modalActiveTab === 'download' && 'Download'}
-              {modalActiveTab === 'palette' && 'palette'}
-            </div>
+          <div className={styles.popupSmall}>
+            <Icon
+              icon={addToCollectionResult?.success ? 'check' : 'close'}
+              color="mediumGradient"
+            />
+            <div>{addToCollectionResult?.message}</div>
+            {addToCollectionResult?.success && (
+              <Button
+                icon="edit"
+                text="Change"
+                small
+                onClick={() => setModalAddToCollectionExtended(true)}
+              />
+            )}
+          </div>
+        </Modal>
+
+        <Modal
+          show={modalAddToCollectionExtended === true}
+          position="bottomRight"
+          size={popupSize}
+          backgroundOverlay={false}
+        >
+          <div className={styles.popupLarge}>
+            <Headline>Add to</Headline>
+            <div>Coll123</div>
+            <Button
+              icon="check"
+              text="Done"
+              onClick={() => setModalAddToCollectionExtended(null)}
+            />
           </div>
         </Modal>
       </main>
-
-      <Modal
-        show={modalAddToCollectionExtended === false}
-        position="bottomRightSlide"
-        size={popupSize}
-        backgroundOverlay={false}
-      >
-        <div className={styles.popupSmall}>
-          <Icon icon={addToCollectionResult?.success ? 'check' : 'close'} color="mediumGradient" />
-          <div>{addToCollectionResult?.message}</div>
-          {addToCollectionResult?.success && (
-            <Button
-              icon="edit"
-              text="Change"
-              small
-              onClick={() => setModalAddToCollectionExtended(true)}
-            />
-          )}
-        </div>
-      </Modal>
-
-      <Modal
-        show={modalAddToCollectionExtended === true}
-        position="bottomRight"
-        size={popupSize}
-        backgroundOverlay={false}
-      >
-        <div className={styles.popupLarge}>
-          <Headline>Add to</Headline>
-          <div>Coll123</div>
-          <Button icon="check" text="Done" onClick={() => setModalAddToCollectionExtended(null)} />
-        </div>
-      </Modal>
     </>
   );
 }
